+++ src/hacks/tps_bypass.cpp
@@ -32,7 +32,8 @@ class $modify(GJBaseGameLayer) {
         
     }
 
-    float getModifiedDelta(float dt) {
+    // 2.2081: getModifiedDelta returns a double
+    double getModifiedDelta(float dt) {
         if (!Global::get().tpsEnabled) return GJBaseGameLayer::getModifiedDelta(dt);
         if (Global::getTPS() == 240.f) return GJBaseGameLayer::getModifiedDelta(dt);
         if (!PlayLayer::get()) return GJBaseGameLayer::getModifiedDelta(dt);