+++ src/ui/button_edit_layer.hpp
@@ -20,7 +20,7 @@ struct MovingButton {
     cocos2d::CCPoint offset = ccp(0, 0);
 };
 
-class ButtonEditLayer : public geode::Popup<> {
+class ButtonEditLayer : public xdbot::Popup<> {
 
 private: