+++ src/ui/load_macro_layer.hpp
@@ -11,8 +11,8 @@ class MacroCell : public CCNode {
 	std::filesystem::path path;
 	std::time_t date;
 
-	geode::Popup<>* menuLayer = nullptr;
-	geode::Popup<>* mergeLayer = nullptr;
+	xdbot::Popup<>* menuLayer = nullptr;
+	xdbot::Popup<>* mergeLayer = nullptr;
 	CCLayer* loadLayer = nullptr;
 
 	bool isMerge = false;
@@ -22,9 +22,9 @@ public:
 	CCMenu* menu = nullptr;
 	CCMenuItemToggler* toggler = nullptr;
 
-	static MacroCell* create(std::filesystem::path path, std::string name, std::time_t date, geode::Popup<>* menuLayer, geode::Popup<>* mergeLayer, CCLayer* loadLayer);
+	static MacroCell* create(std::filesystem::path path, std::string name, std::time_t date, xdbot::Popup<>* menuLayer, xdbot::Popup<>* mergeLayer, CCLayer* loadLayer);
 
-	bool init(std::filesystem::path path, std::string name, std::time_t date, geode::Popup<>* menuLayer, geode::Popup<>* mergeLayer, CCLayer* loadLayer);
+	bool init(std::filesystem::path path, std::string name, std::time_t date, xdbot::Popup<>* menuLayer, xdbot::Popup<>* mergeLayer, CCLayer* loadLayer);
 
 	void onLoad(CCObject*);
 
@@ -39,11 +39,11 @@ public:
 	void selectMacro(bool single);
 };
 
-class LoadMacroLayer : public geode::Popup<geode::Popup<>*, geode::Popup<>*, bool>, public TextInputDelegate {
+class LoadMacroLayer : public xdbot::Popup<xdbot::Popup<>*, xdbot::Popup<>*, bool>, public TextInputDelegate {
 public:
 
-	geode::Popup<>* menuLayer = nullptr;
-	geode::Popup<>* mergeLayer = nullptr;
+	xdbot::Popup<>* menuLayer = nullptr;
+	xdbot::Popup<>* mergeLayer = nullptr;
 	CCMenu* menu = nullptr;
 
 	CCMenuItemToggler* selectAllToggle = nullptr;
@@ -66,11 +66,11 @@ public:
 	bool isMerge = false;
 	bool invertSort = false;
 
-	static LoadMacroLayer* create(geode::Popup<>* layer, geode::Popup<>* layer2, bool autosaves);
+	static LoadMacroLayer* create(xdbot::Popup<>* layer, xdbot::Popup<>* layer2, bool autosaves);
 
-	bool setup(geode::Popup<>* layer, geode::Popup<>* layer2, bool autosaves) override;
+	bool setup(xdbot::Popup<>* layer, xdbot::Popup<>* layer2, bool autosaves) override;
 
-	static void open(geode::Popup<>* layer, geode::Popup<>* layer2, bool autosaves = false);
+	static void open(xdbot::Popup<>* layer, xdbot::Popup<>* layer2, bool autosaves = false);
 
 	void openFolder(CCObject*) {
 		file::openFolder(Mod::get()->getSettingValue<std::filesystem::path>(isAutosaves ? "autosaves_folder" : "macros_folder"));