+++ src/ui/mirror_settings_layer.hpp
@@ -1,7 +1,7 @@
 #include "../includes.hpp"
 #include "record_layer.hpp"
 
-class MirrorSettingsLayer : public geode::Popup<> {
+class MirrorSettingsLayer : public xdbot::Popup<> {
 
 private: