+++ src/ui/noclip_settings_layer.hpp
@@ -1,6 +1,6 @@
 #include "../includes.hpp"
 
-class NoclipSettingsLayer : public geode::Popup<> {
+class NoclipSettingsLayer : public xdbot::Popup<> {
 
 private: